refactored collection.rs into several files
