+++ src/collection.rs
@@ -5,695 +5,12 @@ use std::rc::Rc;
     
 use macros::* ;
 use adapton_sigs::* ;
+use collection_traits::*;
 use quickcheck::Arbitrary;
 use quickcheck::Gen;
 use std::num::Zero;
 
 use rand::{Rng,Rand};
-
-#[derive(Debug,Hash,PartialEq,Eq,Clone,Rand)]
-pub enum CursorEdit<X,Dir> {
-    Insert  (Dir,X), // Inserts an element
-    Remove  (Dir),   // Removes an element (name-oblivious)
-    Replace (Dir,X), // Replace an element (name-oblivious)
-    Goto    (Dir),   // Move the cursor
-}
-
-#[derive(Debug,Hash,PartialEq,Eq,Clone,Rand)]
-pub enum Cmd<X,Dir,Name> {
-    Basic    (CursorEdit<X,Dir>),
-    InsName  (Dir,Name), // Insert name
-    RemName  (Dir),      // Remove immediately-adjacent
-    ShowView (ListReduce),   // Show the given view
-    HideView (ListReduce),   // Remove the given view
-}
-
-#[derive(Debug,Hash,PartialEq,Eq,Clone,Rand)]
-pub enum ListTransf {
-    Sort, Reverse
-}
-
-#[derive(Debug,Hash,PartialEq,Eq,Clone,Rand)]
-pub enum ListReduce {
-    Max, Min, Median,
-    DemandAll(ListTransf),
-    DemandN(ListTransf, usize),
-}
-
-impl Arbitrary for Dir2 {
-    fn arbitrary<G:Gen> (g: &mut G) -> Self {
-        if g.gen() { Dir2::Left  }
-        else       { Dir2::Right }
-    }
-    fn shrink(&self) -> Box<Iterator<Item=Self>> {
-        match *self {
-            Dir2::Right => Box::new(Some(Dir2::Left).into_iter()),
-            Dir2::Left  => Box::new(None.into_iter())
-        }
-    }
-}
-
-impl<X:Arbitrary+Sized+Rand>
-    Arbitrary for CursorEdit<X,Dir2>
-{
-    fn arbitrary<G:Gen> (g: &mut G) -> Self {
-        //match Rand::rand(g) as ListEditCmd<X,Name> {
-        match g.gen_range(0, 100) {
-            0  ... 50  => { CursorEdit::Insert (Arbitrary::arbitrary(g), Arbitrary::arbitrary(g)) },
-            50 ... 60  => { CursorEdit::Remove (Arbitrary::arbitrary(g)) },
-            60 ... 70  => { CursorEdit::Replace(Arbitrary::arbitrary(g), Arbitrary::arbitrary(g)) },
-            70 ... 95 =>  { CursorEdit::Goto   (Dir2::Right) },
-            95 ... 100 => { CursorEdit::Goto   (Dir2::Left)  },
-            _ => unreachable!()
-        }
-    }
-    fn shrink(&self) -> Box<Iterator<Item=Self>> {
-        Box::new(None.into_iter())
-    }
-}
-
-pub trait ExperimentT<A:Adapton,X,Out> {
-    type ListEdit : ListEdit<A,X> ;
-    fn run (&mut A, Vec<CursorEdit<X,Dir2>>, view:ListReduce) -> Vec<(Out,Cnt)> ;
-}
-
-fn has_consecutive_names<A:Adapton,X,L:ListT<A,X>> (st:&mut A, list:L::List) -> bool {
-    L::elim(st, list,
-            |st| false,
-            |st,x,xs| has_consecutive_names::<A,X,L> (st, xs),
-            |st,n,xs|
-            L::elim(st, xs,
-                    |st| false,
-                    |st,y,ys| has_consecutive_names::<A,X,L> (st, ys),
-                    |st,m,ys| true))
-}
-
-pub struct Experiment ;
-impl<A:Adapton,X:Zero+Hash+Debug+PartialEq+Eq+Clone+PartialOrd> ExperimentT<A,X,Vec<X>>
-    for Experiment
-{
-    type ListEdit = ListZipper<A,X,List<A,X>> ;
-    fn run (st:&mut A, edits:Vec<CursorEdit<X,Dir2>>, view:ListReduce) -> Vec<(Vec<X>,Cnt)> {
-        debug!("run");
-        let mut outs : Vec<(Vec<X>,Cnt)> = Vec::new();
-        let mut z : ListZipper<A,X,List<A,X>> = Self::ListEdit::empty(st) ;
-        let mut loop_cnt = 0 as usize;
-        for edit in edits.into_iter() {
-            debug!("\n----------------------- Loop head; count={}", loop_cnt);
-            debug!("zipper: {:?}", z);
-            let consecutive_left  = has_consecutive_names::<A,X,List<A,X>>(st, z.left.clone());
-            let consecutive_right = has_consecutive_names::<A,X,List<A,X>>(st, z.right.clone());
-            debug!("zipper names: consecutive left: {}, consecutive right: {}",
-                     consecutive_left, consecutive_right);
-            //assert!(!consecutive_left);  // Todo-Later: This assertion generally fails for random interactions
-            //assert!(!consecutive_right); // Todo-Later: This assertion generally fails for random interactions
-            debug!("edit:   {:?}", edit);
-            let (out, cnt) = st.cnt(|st|{
-                let z_next = eval_edit::<A,X,Self::ListEdit>(st, edit, z.clone(), loop_cnt);
-                let tree = Self::ListEdit::get_tree::<Tree<A,X,u32>>(st, z_next.clone(), Dir2::Left);
-                debug!("tree:   {:?}", tree);
-                let nm = st.name_of_string("eval_reduce".to_string());
-                let out = st.ns(nm, |st|eval_reduce::<A,X,Tree<A,X,u32>>(st, tree, &view) );
-                z = z_next;
-                loop_cnt = loop_cnt + 1;
-                out
-            }) ;
-            debug!("out:    {:?}", out);
-            debug!("cnt:    {:?}", cnt);
-            outs.push((out,cnt));
-        } outs
-    }
-}
-
-fn eval_edit<A:Adapton,X,E:ListEdit<A,X>> (st:&mut A, edit:CursorEdit<X,E::Dir>, z:E::State, id:usize) -> E::State {
-    match edit {
-        CursorEdit::Insert(dir,x) => {
-            let z = E::clr_names(st, z, dir.clone()) ;
-            let n = A::name_of_usize(st, id);
-            let m = A::name_of_string(st, format!("input{}", id)) ;
-            let z = E::ins_cell(st, z, dir.clone(), m) ;
-            let z = E::ins_name(st, z, dir.clone(), n) ;
-            let z = E::insert(st, z, dir, x) ;
-            z },
-        CursorEdit::Remove(dir) => {
-            let (z, _) = E::remove (st, z, dir.clone()) ;
-            let (z, _) = E::rem_name(st, z, dir) ;
-            z },
-        CursorEdit::Goto(dir) => {
-            let (z, _) = E::goto (st, z, dir.clone()) ;
-            z },
-        CursorEdit::Replace(dir,x) => {
-            let (z, _, _) = E::replace(st, z, dir, x) ;
-            z },
-    }
-}
-
-fn eval_reduce<A:Adapton,X:Zero+Hash+Eq+PartialOrd+Debug+Clone,T:TreeT<A,X>> (st:&mut A, tree:T::Tree, red:&ListReduce) -> Vec<X> {
-    match *red {
-        ListReduce::Max => { let x = tree_reduce_monoid::<A,X,T,_> (st, tree, X::zero(), &|st,x,y| if x > y {x} else {y}) ; vec![ x ] },
-        ListReduce::Min => { let x = tree_reduce_monoid::<A,X,T,_> (st, tree, X::zero(), &|st,x,y| if x < y {x} else {y}) ; vec![ x ] },
-        ListReduce::Median => panic!(""),
-        ListReduce::DemandAll(ListTransf::Sort) => panic!(""),
-        ListReduce::DemandAll(ListTransf::Reverse) => panic!(""),
-        ListReduce::DemandN(_,_) => panic!(""),
-    }
-}
-
-/// `ListEdit<A,X,L>` gives a simple notion of list-editing that is
-/// generic with respect to adapton implementation `A`, list element
-/// type `X`, and list implementation `L`.
-pub trait ListEdit<A:Adapton,X> {
-    /// The State of the Editor is abstract.
-    type State ;
-    /// Lists with foci admit two directions for movement.
-    type Dir:Clone+Hash+Eq+PartialEq=Dir2;
-    fn empty    (&mut A) -> Self::State;
-    fn insert   (&mut A, Self::State, Self::Dir, X) -> Self::State;
-    fn remove   (&mut A, Self::State, Self::Dir)    -> (Self::State, Option<X>);
-    fn replace  (&mut A, Self::State, Self::Dir, X) -> (Self::State, X, bool);
-    fn goto     (&mut A, Self::State, Self::Dir)    -> (Self::State, bool);
-    fn observe  (&mut A, Self::State, Self::Dir)    -> (Self::State, Option<X>);
-
-    // Insert/remove names from the list content:
-    fn clr_names (&mut A, Self::State, Self::Dir) -> Self::State;
-    fn ins_name  (&mut A, Self::State, Self::Dir, A::Name) -> Self::State;
-    fn ins_cell  (&mut A, Self::State, Self::Dir, A::Name) -> Self::State;
-    fn rem_name  (&mut A, Self::State, Self::Dir) -> (Self::State, Option<A::Name>);
-
-    fn get_list<L:ListT<A,X>,T:TreeT<A,X>> (&mut A, Self::State, Self::Dir) -> L::List;
-    fn get_tree<T:TreeT<A,X>>              (&mut A, Self::State, Self::Dir) -> T::Tree;
-
-
-}
-/// Lists are one-dimensional structures; movement admits two possible directions.
-#[derive(Debug,Hash,PartialEq,Eq,Clone)]
-pub enum Dir2 { Left, Right }
-
-impl Rand for Dir2 {
-    fn rand<R: Rng>(rng: &mut R) -> Dir2 {
-        if rng.gen_weighted_bool(2) { Dir2::Left } else { Dir2::Right }
-    }
-}
-
-
-/// Lists with a focus; suitable to implement `ListEdit`.
-#[derive(Debug,Hash,PartialEq,Eq,Clone)]
-pub struct ListZipper<A:Adapton,X,L:ListT<A,X>> {
-    /// Elements to the left of the focus, nearest to furthest.
-    pub left: L::List,
-    /// Elements to the right of the focus, nearest to furthest.
-    pub right: L::List,
-}
-
-/// Implement `ListEdit` for `ListZipper` generically with respect to
-/// adapton implementation `A`, list element type `X`, and list
-/// implementation `L`.
-impl<A:Adapton
-    ,X:Debug+Hash+PartialEq+Eq+Clone
-    ,L:ListT<A,X>
-    >
-    ListEdit<A,X>
-    for
-    ListZipper<A,X,L>
-{
-    type State=ListZipper<A,X,L>;
-    type Dir=Dir2;
-    
-    fn clr_names (st:&mut A, zip:Self::State, dir:Self::Dir) -> Self::State {
-        match dir {
-            Dir2::Left => L::elim_move
-                (st, zip.left, zip.right,
-                 |st,right| ListZipper{left:L::nil(st), right:right},
-                 |st,x,left,right| ListZipper{left:L::cons(st,x,left), right:right},
-                 |st,nm,left,right| {
-                     let right = L::name(st,nm,right);
-                     Self::clr_names(st, ListZipper{left:left, right:right}, dir)}
-                 ),
-            Dir2::Right => L::elim_move
-                (st, zip.right, zip.left,
-                 |st,left| ListZipper{right:L::nil(st), left:left},
-                 |st,x,right,left| ListZipper{right:L::cons(st,x,right), left:left},
-                 |st,nm,right,left| {
-                     let left = L::name(st,nm,left);
-                     Self::clr_names(st, ListZipper{left:left, right:right}, dir)}
-                 ),
-        }
-    }
-
-    fn ins_name (st:&mut A, zip:Self::State, dir:Self::Dir, name:A::Name) -> Self::State {
-        match dir {
-            Dir2::Left =>
-                ListZipper{left:L::name(st, name, zip.left),
-                           right:zip.right},
-            Dir2::Right =>
-                ListZipper{left:zip.left,
-                           right:L::name(st, name, zip.right)},
-        }
-    }
-
-    fn ins_cell (st:&mut A, zip:Self::State, dir:Self::Dir, name:A::Name) -> Self::State {
-        match dir {
-            Dir2::Left => {
-                let art = st.cell(name, zip.left) ;
-                let art = st.read_only(art);
-                ListZipper{left:L::art(st, art),
-                           right:zip.right}},                
-            Dir2::Right => {
-                let art = st.cell(name, zip.right) ;
-                let art = st.read_only(art);
-                ListZipper{left:zip.left,
-                           right:L::art(st, art)}},
-        }
-    }
-
-    fn rem_name  (st:&mut A, zip:Self::State, dir:Self::Dir) -> (Self::State, Option<A::Name>) {
-        match dir {
-            Dir2::Left => L::elim_move
-                (st, zip.left, zip.right,
-                 |st,right|         (ListZipper{left:L::nil(st), right:right}, None ),
-                 |_,x,left,right|   (ListZipper{left:left,       right:right}, None ),
-                 |st,nm,left,right| (ListZipper{left:left, right:right},       Some(nm))
-                 ),
-            Dir2::Right => L::elim_move
-                (st, zip.right, zip.left,
-                 |st,left|          (ListZipper{left:left, right:L::nil(st)}, None ),
-                 |_,x,right,left|   (ListZipper{left:left, right:right},      None ),
-                 |st,nm,right,left| (ListZipper{left:left, right:right},      Some(nm))
-                 ),
-        }
-    }
-    
-    fn empty (st: &mut A) -> Self::State {
-        let nil1 = L::nil(st);
-        let nil2 = nil1.clone();
-        ListZipper{left:nil1, right:nil2}
-    }
-    
-    fn insert (st:&mut A, zip:Self::State, dir:Self::Dir, x:X) -> Self::State {
-        match dir {
-            Dir2::Left =>
-                ListZipper{left:L::cons(st, x, zip.left),
-                           right:zip.right},
-            Dir2::Right =>
-                ListZipper{left:zip.left,
-                           right:L::cons(st, x, zip.right)},
-        }
-    }
-
-    fn remove  (st:&mut A, zip:Self::State, dir:Self::Dir) -> (Self::State, Option<X>) {
-        match dir {
-            Dir2::Left => L::elim_move
-                (st, zip.left, zip.right,
-                 |st,right|         (ListZipper{left:L::nil(st), right:right}, None   ),
-                 |_,x,left,right|   (ListZipper{left:left,       right:right}, Some(x)),
-                 |st,nm,left,right| {let zip = ListZipper{left:left, right:L::name(st,nm,right)};
-                                     Self::remove (st, zip, Dir2::Left)}
-                 ),
-            Dir2::Right => L::elim_move
-                (st, zip.right, zip.left,
-                 |st,left|          (ListZipper{left:left, right:L::nil(st)}, None   ),
-                 |_,x,right,left|   (ListZipper{left:left, right:right},      Some(x)),
-                 |st,nm,right,left| {let zip = ListZipper{left:L::name(st,nm,left), right:right};
-                                     Self::remove (st, zip, Dir2::Right)}
-                 ),
-        }
-    }
-
-    
-    fn goto (st:&mut A, zip:Self::State, dir:Self::Dir) -> (Self::State, bool) {
-        match dir {
-            Dir2::Left => L::elim_move
-                (st, zip.left, zip.right,
-                 |st,right|         (ListZipper{left:L::nil(st), right:right}              , false ),
-                 |st,x,left,right|  (ListZipper{left:left,       right:L::cons(st,x,right)}, true  ),
-                 |st,nm,left,right| {let zip = ListZipper{left:left, right:L::name(st,nm,right)};
-                                     Self::goto (st, zip, Dir2::Left)}
-                 ),
-            Dir2::Right => L::elim_move
-                (st, zip.right, zip.left,
-                 |st,left|          (ListZipper{left:left,              right:L::nil(st)}, false ),
-                 |st,x,right,left|  (ListZipper{left:L::cons(st,x,left),right:right}     , true  ),
-                 |st,nm,right,left| {let zip = ListZipper{left:L::name(st,nm,left), right:right};
-                                     Self::goto (st, zip, Dir2::Right)}
-                 ),
-        }
-    }
-
-    fn observe (st:&mut A, zip:Self::State, dir:Self::Dir) -> (Self::State,Option<X>) {
-        match dir {
-            Dir2::Left => L::elim_move
-                (st, zip.left, zip.right,
-                 |st,right|         (ListZipper{left:L::nil(st), right:right}, None),
-                 |st,x,left,right| {let x2 = x.clone();
-                                    (ListZipper{left:L::cons(st,x,left), right:right}, Some(x2))},
-                 |st,nm,left,right|{let zip = ListZipper{left:left,right:L::name(st,nm,right)};
-                                    Self::observe (st, zip, Dir2::Left)}
-                 ),
-            Dir2::Right => L::elim_move
-                (st, zip.right, zip.left,
-                 |st,left|         (ListZipper{left:left, right:L::nil(st)}, None),
-                 |st,x,right,left| {let x2 = x.clone();
-                                    (ListZipper{left:left, right:L::cons(st,x,right)}, Some(x2))},
-                 |st,nm,right,left|{let zip = ListZipper{left:L::name(st,nm,left),right:right};
-                                    Self::observe (st, zip, Dir2::Right)}
-                 ),
-        }
-    }
-
-    fn replace (st:&mut A, zip:Self::State, dir:Self::Dir, y:X) -> (Self::State, X, bool) {
-        match dir {
-            Dir2::Left => L::elim_move
-                (st, zip.left, (zip.right, y),
-                 |st,(right,y)|        (ListZipper{left:L::nil(st),         right:right}, y, false),
-                 |st,x,left,(right,y)| (ListZipper{left:L::cons(st,y,left), right:right}, x, true ),
-                 |st,nm,left,(right,y)|{let zip = ListZipper{left:left,right:L::name(st,nm,right)};
-                                        Self::replace (st, zip, Dir2::Left, y)}
-                 ),
-            Dir2::Right => L::elim_move
-                (st, zip.right, (zip.left,y),
-                 |st,(left,y)|         (ListZipper{left:left, right:L::nil(st)},          y, false),
-                 |st,x,right,(left,y)| (ListZipper{left:left, right:L::cons(st,y,right)}, x, true ),
-                 |st,nm,right,(left,y)|{let zip = ListZipper{left:L::name(st,nm,left),right:right};
-                                        Self::replace (st, zip, Dir2::Right, y)}
-                 ),
-        }
-    }
-
-    fn get_list<N:ListT<A,X>,T:TreeT<A,X>>
-        (st:&mut A, zip:Self::State, dir:Self::Dir) -> N::List
-    {
-        let tree = Self::get_tree::<T>(st, zip, dir);
-        list_of_tree::<A,X,N,T>(st, tree)
-    }
-
-    /// Creates a tree whose leaves hold the contents of the zipper, in order.
-    /// When `dir=Left`,  the tree's leaves are ordered from left-to-right, i.e., as (rev left) @ right.
-    /// When `dir=Right`, the tree's leaves are ordered from right-to-left, i.e., as (rev right) @ left.
-    fn get_tree<T:TreeT<A,X>>
-        (st:&mut A, zip:Self::State, dir:Self::Dir) -> T::Tree
-    {
-        match dir {
-            Dir2::Left  => {
-                let left  = tree_of_list::<A,X,T,L>(st, Dir2::Right, zip.left);
-                let right = tree_of_list::<A,X,T,L>(st, Dir2::Left,  zip.right);
-                tree_append::<A,X,T>(st, left, right)}
-            
-            Dir2::Right => {
-                let right = tree_of_list::<A,X,T,L>(st, Dir2::Right, zip.right);
-                let left  = tree_of_list::<A,X,T,L>(st, Dir2::Left,  zip.left);
-                tree_append::<A,X,T>(st, right, left)}
-        }
-    }
-}
-
-
-pub trait ListT<A:Adapton,Hd> : Debug+Clone {
-    type List : Debug+Hash+PartialEq+Eq+Clone ;
-    
-    fn nil  (&mut A) -> Self::List ;
-    fn cons (&mut A, Hd, Self::List) -> Self::List ;
-    
-    // requisite "adaptonic" constructors: `name` and `art`:
-    fn name (&mut A, A::Name, Self::List) -> Self::List ;
-    fn art  (&mut A, Art<Self::List,A::Loc>) -> Self::List ;
-    fn rc   (&mut A, Rc<Self::List>) -> Self::List ;
-
-    fn elim<Res,Nil,Cons,Name> (&mut A, Self::List, Nil, Cons, Name) -> Res
-        where Nil:FnOnce(&mut A) -> Res
-        ,    Cons:FnOnce(&mut A, Hd, Self::List) -> Res
-        ,    Name:FnOnce(&mut A, A::Name, Self::List) -> Res ;
-
-    fn elim_move<Arg,Res,Nil,Cons,Name> (&mut A, Self::List, Arg, Nil, Cons, Name) -> Res
-        where Nil:FnOnce(&mut A, Arg) -> Res
-        ,    Cons:FnOnce(&mut A, Hd, Self::List, Arg) -> Res
-        ,    Name:FnOnce(&mut A, A::Name, Self::List, Arg) -> Res ;
-
-    // Derived from `cons` and `nil` above:
-    fn singleton (st:&mut A, hd:Hd) -> Self::List {
-        let nil = Self::nil(st);
-        Self::cons(st, hd, nil)
-    }
-
-    // Derived from `elim` above:
-    fn is_empty (st:&mut A, list:Self::List) -> bool {
-        Self::elim(st, list,
-                   |_|       true,
-                   |_,_,_|   false,
-                   |st,_,tl| Self::is_empty(st,tl))
-    }
-}
-
-pub trait TreeT<A:Adapton,Leaf> {
-    type Lev  : Debug+Hash+PartialEq+Eq+Clone ;
-    type Tree : Debug+Hash+PartialEq+Eq+Clone ;
-
-    fn lev<X:Hash>(&X) -> Self::Lev ;
-    fn lev_bits () -> Self::Lev ;
-    fn lev_zero () -> Self::Lev ;
-    fn lev_inc (Self::Lev) -> Self::Lev ;
-    fn lev_max () -> Self::Lev ;
-    fn lev_add (Self::Lev, Self::Lev) -> Self::Lev ;
-    fn lev_lte (Self::Lev, Self::Lev) -> bool ;
-    
-    fn nil  (&mut A) -> Self::Tree ;
-    fn leaf (&mut A, Leaf) -> Self::Tree ;
-    fn bin  (&mut A, Self::Lev, Self::Tree, Self::Tree) -> Self::Tree ;
-
-    // requisite "adaptonic" constructors: `name` and `art`:
-    fn name (&mut A, A::Name, Self::Lev, Self::Tree, Self::Tree) -> Self::Tree ;
-    fn art  (&mut A, Art<Self::Tree,A::Loc>) -> Self::Tree ;
-    fn rc   (&mut A, Rc<Self::Tree>) -> Self::Tree ;
-
-    fn elim<Res,NilC,LeafC,BinC,NameC>
-        (&mut A, Self::Tree, NilC, LeafC, BinC, NameC) -> Res
-        where NilC  : FnOnce(&mut A) -> Res
-        ,     LeafC : FnOnce(&mut A, Leaf) -> Res
-        ,     BinC  : FnOnce(&mut A, Self::Lev,  Self::Tree, Self::Tree) -> Res
-        ,     NameC : FnOnce(&mut A, A::Name, Self::Lev, Self::Tree, Self::Tree) -> Res
-        ;
-
-    fn elim_move<Arg,Res,NilC,LeafC,BinC,NameC>
-        (&mut A, Self::Tree, Arg, NilC, LeafC, BinC, NameC) -> Res
-        where NilC  : FnOnce(&mut A, Arg) -> Res
-        ,     LeafC : FnOnce(&mut A, Leaf, Arg) -> Res
-        ,     BinC  : FnOnce(&mut A, Self::Lev,  Self::Tree, Self::Tree, Arg) -> Res
-        ,     NameC : FnOnce(&mut A, A::Name, Self::Lev, Self::Tree, Self::Tree, Arg) -> Res
-        ;
-    
-    fn fold_lr<Res:Hash+Debug+Eq+Clone,LeafC,BinC,NameC>
-        (st:&mut A, tree:Self::Tree, res:Res, leaf:&LeafC, bin:&BinC, name:&NameC) -> Res
-        where LeafC:Fn(&mut A, Leaf,    Res ) -> Res
-        ,      BinC:Fn(&mut A, Self::Lev,     Res ) -> Res 
-        ,     NameC:Fn(&mut A, A::Name, Self::Lev, Res ) -> Res 
-    {
-        Self::elim_move
-            (st, tree, res,
-             |_,res| res,
-             |st,x,res| leaf(st, x, res),
-             |st,x,l,r,res| {
-                 let res = Self::fold_lr(st, l, res, leaf, bin, name);
-                 let res = bin(st, x, res);
-                 let res = Self::fold_lr(st, r, res, leaf, bin, name);
-                 res
-             },
-             |st,n,x,l,r,res| {
-                 let (n1,n2,n3) = st.name_fork3(n.clone());
-                 let res = memo!(st, n1 =>> Self::fold_lr, tree:l, res:res ;; leaf:leaf, bin:bin, name:name);
-                 let res = name(st, n2, x, res);
-                 let res = memo!(st, n3 =>> Self::fold_lr, tree:r, res:res ;; leaf:leaf, bin:bin, name:name);
-                 res
-             }
-             )
-    }
-
-    fn fold_rl<Res:Hash+Debug+Eq+Clone,LeafC,BinC,NameC>
-        (st:&mut A, tree:Self::Tree, res:Res, leaf:&LeafC, bin:&BinC, name:&NameC) -> Res
-        where LeafC:Fn(&mut A, Leaf,    Res ) -> Res
-        ,      BinC:Fn(&mut A, Self::Lev,     Res ) -> Res
-        ,     NameC:Fn(&mut A, A::Name, Self::Lev, Res ) -> Res
-    {
-        Self::elim_move
-            (st, tree, res,
-             |_,res| res,
-             |st,x,res| leaf(st, x, res),
-             |st,x,l,r,res| {
-                 let res = Self::fold_rl(st, r, res, leaf, bin, name);
-                 let res = bin(st, x, res);
-                 let res = Self::fold_rl(st, l, res, leaf, bin, name);
-                 res
-             },
-             |st,n,x,l,r,res| {
-                 let (n1,n2,n3) = st.name_fork3(n);
-                 let res = memo!(st, n1 =>> Self::fold_rl, tree:r, res:res ;; leaf:leaf, bin:bin, name:name);
-                 let res = name(st, n2, x, res);
-                 let res = memo!(st, n3 =>> Self::fold_rl, tree:l, res:res ;; leaf:leaf, bin:bin, name:name);
-                 res
-             }
-             )
-    }
-
-    fn fold_up<Res:Hash+Debug+Eq+Clone,NilC,LeafC,BinC,NameC>
-        (st:&mut A, tree:Self::Tree, nil:&NilC, leaf:&LeafC, bin:&BinC, name:&NameC) -> Res
-        where  NilC:Fn(&mut A) -> Res
-        ,     LeafC:Fn(&mut A, Leaf              ) -> Res
-        ,      BinC:Fn(&mut A, Self::Lev,     Res, Res ) -> Res
-        ,     NameC:Fn(&mut A, A::Name, Self::Lev, Res, Res ) -> Res
-    {
-        Self::elim
-            (st, tree,
-             |st| nil(st),
-             |st,x| leaf(st, x),
-             |st,x,l,r| {
-                 let resl = Self::fold_up(st, l, nil, leaf, bin, name);
-                 let resr = Self::fold_up(st, r, nil, leaf, bin, name);
-                 let res = bin(st, x, resl, resr);
-                 res
-             },
-             |st,n,x,l,r| {
-                 let (n1,n2,n3) = st.name_fork3(n);
-                 let resl = memo!(st, n1 =>> Self::fold_up, tree:l ;; nil:nil, leaf:leaf, bin:bin, name:name);
-                 let resr = memo!(st, n2 =>> Self::fold_up, tree:r ;; nil:nil, leaf:leaf, bin:bin, name:name);
-                 let res = name(st, n3, x, resl, resr);
-                 res
-             }
-             )
-    }
-}
-
-pub fn tree_append
-    < A:Adapton
-    , X:Hash+Clone+Debug
-    , T:TreeT<A,X>
-    >
-    (st:&mut A,tree1:T::Tree,tree2:T::Tree) -> T::Tree
-{
-    // XXX: This is a hack. Make this balanced, a la Pugh 1989.
-    T::bin(st, T::lev_max(), tree1, tree2)
-}
-
-// pub fn tree_append
-//     <A:Adapton
-//     ,X:Clone+Hash+Eq+Debug
-//     ,T:TreeT<A,X,()>
-//     >
-//     (st:&mut A, tree1:T::Tree, tree2:T::Tree) -> T::Tree
-// {
-//     T::elim_move
-//         (st, tree1, tree2,         
-//          /* Nil */  |st,       tree2| tree2,
-//          /* Leaf */ |st,leaf,  tree2| {
-//              T::elim_move
-//                  (st, tree2, leaf,
-//                   /* Nil */  |st, leaf1| T::leaf(st,leaf1),
-//                   /* Leaf */ |st, leaf2, leaf1| {
-//                       let l1 = T::leaf(st,leaf1);
-//                       let l2 = T::leaf(st,leaf2);
-//                       T::bin(st, (), l1, l2)
-//                   },
-//                   /* Bin */  |st, l2, r2, leaf1| {
-//                       let leaf = T::leaf(st, leaf);
-//                       /*T::bin(st, (), leaf, tree2)*/
-//                       panic!("TODO")
-//                   },
-//                   /* Name */ |st, nm, l2, r2, leaf1| {
-//                       panic!("TODO")
-//                   })
-//          },
-//          /* Bin */ |st,_,l,r, tree2| {
-//              /*
-//              T::elim_move
-//                  (st, tree2, (l,r),
-//                   /* Nil */ |st, 
-                  
-//                   let tree1 = T::bin(st, (), l, r);
-//              T::bin(st, (), tree1, tree2)
-//                  */
-//                  panic!("")
-//          },
-//          /* Name */ |st,_,l,r, tree2| panic!(""),
-//          )
-// }
-
-pub fn tree_reduce_monoid<A:Adapton,Elm:Eq+Hash+Clone+Debug,T:TreeT<A,Elm>,BinOp>
-    (st:&mut A, tree:T::Tree, zero:Elm, binop:&BinOp) -> Elm
-    where BinOp:Fn(&mut A, Elm, Elm) -> Elm
-{
-    T::fold_up(st, tree,
-                        &|_| zero.clone(),
-                   &|_,leaf| leaf,
-                 &|st,_,l,r| binop(st,l,r),
-               &|st,_,_,l,r| binop(st,l,r),
-               )
-}
-
-pub fn list_reduce_monoid<A:Adapton,Elm:Eq+Hash+Clone+Debug,L:ListT<A,Elm>,BinOp,T:TreeT<A,Elm>>
-    (st:&mut A, list:L::List, zero:Elm, binop:&BinOp) -> Elm
-    where BinOp:Fn(&mut A, Elm, Elm) -> Elm
-{
-    let tree = tree_of_list::<A,Elm,T,L>(st, Dir2::Left, list);
-    tree_reduce_monoid::<A,Elm,T,BinOp>(st, tree, zero, binop)
-}
-
-// pub fn tree_map<A:Adapton,X:Hash+Clone,T:TreeT<A,X>,FX:Hash+Clone,GY:Hash+Clone,FGT:TreeT<A,FX,GY>,F,G>
-//     (st:&mut A, tree:T::Tree, f:&F, g:&G) -> FGT::Tree
-//     where F:Fn(&mut A, X) -> FX
-//     ,     G:Fn(&mut A, Y) -> GY
-// {
-//     T::fold_up(st, tree,
-//                &|st|       FGT::nil(st),
-//                &|st,x|     {let fx = f(st,x);  FGT::leaf(st, fx)},
-//                &|st,y,l,r| {let gy = g(st, y); FGT::bin(st, gy, l, r)},
-//                &|st,n,l,r| FGT::name(st, n, l, r)
-//                )
-// }
-
-pub fn tree_filter<A:Adapton,X:Hash+Clone,T:TreeT<A,X>,F>
-    (st:&mut A, tree:T::Tree, f:&F) -> T::Tree
-    where F:Fn(&mut A, &X) -> bool
-{
-    T::fold_up(st, tree,
-               &|st| T::nil(st),
-               &|st,x| { let fx = f(st,&x);
-                         if fx { T::leaf(st, x) }
-                         else  { T::nil(st) } },
-               &|st,lev,l,r| T::bin(st, lev, l, r),
-               &|st,n,lev,l,r| T::name(st, n, lev, l, r)
-               )
-}
-
-pub fn list_of_tree<A:Adapton,X:Hash+Clone,L:ListT<A,X>,T:TreeT<A,X>>
-    (st:&mut A, tree:T::Tree) -> L::List
-{
-    let nil = L::nil(st);
-    T::fold_rl(st, tree, nil,
-               &|st,x,xs| L::cons(st,x,xs),
-               &|_,_,xs| xs,
-               &|st,n,_,xs| L::name(st,n,xs)
-               )
-}
-
-pub fn rev_list_of_tree<A:Adapton,X:Hash+Clone,L:ListT<A,X>,T:TreeT<A,X>>
-    (st:&mut A, tree:T::Tree) -> L::List
-{
-    let nil = L::nil(st);
-    T::fold_lr(st, tree, nil,
-               &|st,x,xs| L::cons(st,x,xs),
-               &|_ ,_,xs| xs,
-               &|st,n,_,xs| L::name(st,n,xs)
-               )    
-}
-
-pub fn list_of_vec<A:Adapton,X:Clone,L:ListT<A,X>> (st:&mut A, v:Vec<X>) -> L::List {
-    let mut l = L::nil(st);
-    for x in v.iter().rev() { l = L::cons(st,x.clone(),l) }
-    return l
-}
-
-pub fn rev_list_of_vec<A:Adapton,X:Clone,L:ListT<A,X>> (st:&mut A, v:Vec<X>) -> L::List {
-    let mut l = L::nil(st);
-    for x in v.iter() { l = L::cons(st,x.clone(), l) }
-    return l
-}
 
 #[derive(Debug,PartialEq,Eq,Hash,Clone)]
 pub enum List<A:Adapton,Hd> {
@@ -838,241 +155,4 @@ impl
             }
         }
     }
-}
-
-pub fn tree_of_list
-    < A:Adapton
-    , X:Hash+Clone+Debug
-    , T:TreeT<A,X>
-    , L:ListT<A,X>
-    >
-    (st:&mut A, dir_list:Dir2, list:L::List) -> T::Tree {
-        let tnil = T::nil(st);
-        let lnil = L::nil(st);
-        let (tree, list) = tree_of_list_rec::<A,X,T,L>(st, dir_list, list, tnil, T::lev_zero(), T::lev_max());
-        assert_eq!(list, lnil);
-        tree
-    }
-
-pub fn tree_of_list_rec
-    < A:Adapton
-    , X:Hash+Clone+Debug
-    , T:TreeT<A,X>
-    , L:ListT<A,X>
-    >
-    (st:&mut A,
-     dir_list:Dir2, list:L::List,
-     tree:T::Tree, tree_lev:T::Lev, parent_lev:T::Lev)
-     -> (T::Tree, L::List)
-{
-    L::elim_move (
-        st, list, (dir_list, tree, tree_lev, parent_lev),
-
-        /* Nil */
-        |st, (_dir_list, tree, _, _)| (tree, L::nil(st)),
-
-        /* Cons */
-        |st, hd, rest, (dir_list, tree, tree_lev, parent_lev)| {
-            let lev_hd = T::lev_inc ( T::lev(&hd) ) ;
-            if T::lev_lte ( tree_lev , lev_hd.clone() ) && T::lev_lte ( lev_hd.clone() , parent_lev.clone() ) {
-                let leaf = T::leaf(st, hd) ;
-                let (tree2, rest2) = {
-                    tree_of_list_rec::<A,X,T,L> ( st, dir_list.clone(), rest, leaf, T::lev_zero(), lev_hd.clone() )
-                };
-                let tree3 = match dir_list.clone() {
-                    Dir2::Left  => T::bin ( st, lev_hd.clone(), tree,  tree2 ),
-                    Dir2::Right => T::bin ( st, lev_hd.clone(), tree2, tree  ),
-                } ;
-                tree_of_list_rec::<A,X,T,L> ( st, dir_list, rest2, tree3, lev_hd, parent_lev )
-            }
-            else {
-                (tree, L::cons(st,hd,rest))
-            }},
-
-        /* Name */
-        |st, nm, rest, (dir_list, tree, tree_lev, parent_lev)|{
-            let lev_nm = T::lev_inc( T::lev_add( T::lev_bits() , T::lev(&nm) ) ) ;
-            if T::lev_lte ( tree_lev , lev_nm.clone() ) && T::lev_lte ( lev_nm.clone() ,  parent_lev.clone() ) {
-                let nil = T::nil(st) ;
-                let (nm1, nm2, nm3, nm4) = st.name_fork4(nm.clone());
-                let (_, (tree2, rest)) =
-                    eager!(st, nm1 =>> tree_of_list_rec::<A,X,T,L>,
-                           dir_list:dir_list.clone(), list:rest,
-                           tree:nil, tree_lev:T::lev_zero(), parent_lev:lev_nm.clone() ) ;
-                let tree3 = match dir_list.clone() {
-                    Dir2::Left  => T::name ( st, nm.clone(), lev_nm.clone(), tree,  tree2 ),
-                    Dir2::Right => T::name ( st, nm.clone(), lev_nm.clone(), tree2, tree  ),
-                } ;
-                let art = st.cell(nm3, tree3) ;
-                let art = st.read_only( art ) ;
-                let tree3 = T::art( st, art ) ;                
-                let (_, (tree, rest)) =
-                    eager!(st, nm2 =>> tree_of_list_rec::<A,X,T,L>,
-                           dir_list:dir_list.clone(), list:rest,
-                           tree:tree3, tree_lev:lev_nm, parent_lev:parent_lev ) ;
-                let art = st.cell(nm4, tree) ;
-                let art = st.read_only( art ) ;
-                let tree = T::art( st, art ) ;
-                (tree, rest)
-            }
-            else {
-                (tree, L::name(st,nm,rest))
-            }},
-        )
-}
-
-pub fn list_merge<A:Adapton,X:Ord+Clone+Debug,L:ListT<A,X>>
-    (st:&mut A,
-     n1:Option<A::Name>, l1:L::List,
-     n2:Option<A::Name>, l2:L::List  ) -> L::List
-{
-    L::elim_move
-        (st, l1, (n1,n2,l2),
-         /* Nil */  |_, (_, _, l2)| l2,
-         /* Cons */ |st,h1,t1,(n1,n2,l2)|
-         L::elim_move
-         (st, l2, (h1,t1,n1,n2),
-          /* Nil */  |st, (h1, t1, _, _ )| L::cons(st, h1,t1),
-          /* Cons */ |st, h2, t2, (h1, t1, n1, n2)| {
-          if &h1 <= &h2 {
-              let l2 = L::cons(st,h2,t2);
-              match n1 {
-                  None => {
-                      let rest = list_merge::<A,X,L>(st, None, t1, n2, l2);
-                      L::cons(st, h1, rest)
-                  }
-                  Some(n1) => {
-                      let (n1a, n1b) = st.name_fork(n1);
-                      let rest = thunk!(st, n1a =>>
-                                        list_merge::<A,X,L>,
-                                        n1:None, l1:t1, n2:n2, l2:l2);
-                      let rest = L::art(st, rest);
-                      let rest = L::cons(st, h1, rest);
-                      L::name(st, n1b, rest)
-                  }
-              }
-          }
-          else {
-              let l1 = L::cons(st,h1,t1);
-              match n2 {
-                  None => {
-                      let rest = list_merge::<A,X,L>(st, n1, l1, None, t2);
-                      let l = L::cons(st, h2, rest);
-                      l
-                  }
-                  Some(n2) => {
-                      let (n2a, n2b) = st.name_fork(n2);
-                      let rest = thunk!(st, n2a =>>
-                                        list_merge::<A,X,L>,
-                                        n1:n1, l1:l1, n2:None, l2:t2);
-                      let rest = L::art(st, rest);
-                      let rest = L::cons(st, h2, rest);
-                      L::name(st, n2b, rest)
-                  }
-              }
-          }},
-          |st,m2,t2,(h1,t1,n1,_n2)| {
-              let l1 = L::cons(st,h1,t1);
-              list_merge::<A,X,L>(st, n1, l1, Some(m2), t2)
-          }
-          ),
-         |st,m1,t1,(_n1,n2,l2)| {
-             list_merge::<A,X,L>(st, Some(m1), t1, n2, l2)
-         }
-         )
-}
-
-pub fn list_merge_sort<A:Adapton,X:Ord+Hash+Debug+Clone,L:ListT<A,X>,T:TreeT<A,X>>
-    (st:&mut A, list:L::List) -> L::List
-{
-    let tree = tree_of_list::<A,X,T,L>(st, Dir2::Left, list);
-    T::fold_up (st, tree,
-                &|st|                 L::nil(st), 
-                &|st, x|              L::singleton(st, x),
-                &|st, _, left, right| { list_merge::<A,X,L>(st, None, left, None, right) },
-                &|st, n, _, left, right| { let (n1,n2) = st.name_fork(n);
-                                           list_merge::<A,X,L>(st, Some(n1), left, Some(n2), right) },
-                )
-}
-
-pub trait SetT
-    <A:Adapton
-    ,Elm:Hash+Clone>
-{
-    type Set;
-
-    // Intro forms:
-    fn empty(st:&mut A);
-
-    fn update_elm<Update,Res>
-        (st:&mut A, set:Self::Set, elm:Elm,
-         update:&Update)
-         -> (Self::Set, Res)
-        where Update:FnOnce(&mut A, Option<Elm>, Elm) -> (Option<Elm>, Res)
-        ;
-    
-    // Elim forms:
-    fn elim_move<Arg,Res,NilC,ElmC,UnionC>
-        (st:&mut A, set:Self::Set, arg:Arg,
-         nil:&NilC, elm:&ElmC, union:&UnionC)
-         -> Res
-        where NilC:FnOnce(&mut A, Arg) -> Res
-        ,     ElmC:FnOnce(&mut A, Elm, Arg) -> Res
-        ,   UnionC:FnOnce(&mut A, Self::Set, Self::Set, Arg) -> Res
-        ;
-
-    fn is_mem(st:&mut A, set:Self::Set, elm:Elm) -> bool; // TODO: Write based on update, above.
-    fn is_empty(st:&mut A, set:Self::Set) -> bool; // TODO: Write based on elim_with, above.
-}
-
-pub trait MapT
-    <A:Adapton
-    ,Dom:Hash+Clone
-    ,Cod:Hash+Clone>
-{
-    type Map;
-    
-    // Intro forms:
-    fn empty(st:&mut A);
-
-    fn update_pt<Update,Res>
-        (st:&mut A, map:Self::Map,
-         pt:Dom, cod:Option<Cod>, update:&Update) -> (Self::Map, Res)
-        where Update:FnOnce(&mut A, Option<Cod>, Option<Cod>) -> (Option<Cod>, Res)
-        ;
-
-    // General elim form:
-    fn elim_move<Arg,Res,NilC,ElmC,UnionC>
-        (st:&mut A, map:Self::Map, arg:Arg,
-         nil:&NilC, elm:&ElmC, union:&UnionC)
-        where NilC:FnOnce(&mut A, Arg) -> Res
-        ,     ElmC:FnOnce(&mut A, Dom, Option<Cod>) -> Res
-        ,   UnionC:FnOnce(&mut A, Self::Map, Self::Map, Arg) -> Res
-        ;
-    
-    // Special, convenient update_pts:
-    fn rem(st:&mut A, map:Self::Map, x:Dom) -> (Option<Cod>, Self::Map); // TODO: Implement with update_pt
-    fn get(st:&mut A, set:Self::Map, x:Dom) -> Option<Cod>; // TODO: Implement with update_pt
-}
-
-pub trait GraphT
-    <A:Adapton
-    ,Node:Hash+Eq+Clone+Debug
-    ,NodeSet:SetT<A,Node>+Clone+Hash
-    ,NodeMap:MapT<A,Node,NodeSet>>
-{
-    type Graph : Hash+Eq+Clone+Debug;
-
-    // Intro forms:
-    fn empty(st:&mut A);
-    fn add_node(st:&mut A, graph:Self::Graph, node:Node) -> Self::Graph;
-    fn add_edge(st:&mut A, graph:Self::Graph, src:Node, tgt:Node) -> Self::Graph;
-    fn add_succs(st: &mut A, graph:Self::Graph, node:Node, succs:NodeSet) -> Self::Graph;
-
-    // Query forms:
-    fn get_succs(st: &mut A, graph:Self::Graph, node:Node) -> NodeSet;
-    
-    // Other forms:
-    // fn rem_node(st:&mut A, graph:Self::Graph, node:NodeLab) -> Self::Graph;
-    // fn rem_edge(st:&mut A, graph:Self::Graph, edge:Self::Edge) -> Self::Graph;    
 }
